Add ActivityHistoryViewModel and connect it with lib and the fragment.
Through the ActivityHistoryViewModule LiveData is provided to ActivityHistoryFragment,
which makes use of the library through the injected StandardProfile.

The RecyclerViewAdapter has been changed to work with List<FinishedActivity> instead of three List<String> as previously done.

Slight modification to Fitness and StandardProfile in the return vaules of some methods, were required to return a List<> instead of an array, to better fit LiveData.

On Click of an entry in the recycler view, a new entry is added, with random date.
The code is only used to check whether the view updates if an entry is added.